+++ src/Appointments/Appointments/Application/AppointmentCreator.cs
@@ -9,7 +9,7 @@ namespace Appointments.Application{
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
         }
 
-        public async Task Create(Guid appointmentID, DateTime startDateTime, DateTime endDateTime, double duration, string message){
+        public async Task Create(Guid appointmentID, DateTime startDateTime, DateTime endDateTime, int duration, string message){
             await _repository.Create(new Appointment(appointmentID, startDateTime,endDateTime,duration,message));
         }
